requirements.txt
openai==1.14.2
Pillow==10.2.0
pydub==0.25.1
pyttsx3==2.90
streamlit==1.32.2
load_dotenv==1.0.1